javaspring-buoi38

+++ src/main/java/vn/scrip/buoi38_bvn/entites/Category.java
@@ -1,20 +1,19 @@
 package vn.scrip.buoi38_bvn.entites;
 
 import jakarta.persistence.*;
+
 @Entity
 @Table(name = "categories")
 public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
+    private Long id;
     private String name;
 
-    // Getter Setter
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
-
+    // Getters & Setters
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
 }
