+++ src/main/java/vn/scrip/buoi38_bvn/services/BookService.java
@@ -8,16 +8,14 @@ import java.util.List;
 
 @Service
 public class BookService {
-
     private final BookRepository repo;
 
-    public BookService(BookRepository repo) { this.repo = repo; }
+    public BookService(BookRepository repo) {
+        this.repo = repo;
+    }
 
     public List<Book> getAll() { return repo.findAll(); }
-
-    public Book getById(Integer id) { return repo.findById(id).orElse(null); }
-
+    public Book getById(Long id) { return repo.findById(id).orElse(null); }
     public Book save(Book book) { return repo.save(book); }
-
-    public void delete(Integer id) { repo.deleteById(id); }
+    public void delete(Long id) { repo.deleteById(id); }
 }
